% Auto-generated: do not edit by hand
\name{dangerouslySetInnerHtml}

\alias{dangerouslySetInnerHtml}

\title{DangerouslySetInnerHtml component}

\description{
ExampleComponent is an example component. It renders an input with the property `value` which is editable by the user.
}

\usage{
dangerouslySetInnerHtml(id=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. The html content of the div.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

